Correct User#42

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -14,8 +14,7 @@ import java.time.LocalDate;
 
 public class Film {
 
-    private Integer id;
-
+    private long id;
     private String name;
     @Size(min = 1, max = 200)
     private String description;
